Added but not using turn rate to storms and player.
The idea behind adding this is that it will prevent a single orbit for
each spherical object. However, this results in somewhat more chaotic
movement and not a noticeable increase in fun or engagement. if we want to
change these trajectories, I'm considering making it more of a tangible
gameplay mechanic : winds that blow things around or clouds/storms moving
each other with their own force of wind. Perhaps even the player would
also contribute to this as they pass by things.